Start modeling a name-based registry for chips.
A hardware description language needs a set of built-in chips to create
new chips, and a way to store and identify all those chips to create
even more chips. That's the purpose of the registry. Each chip, whether
built-in or composite, is uniquely identified in a registry by its
`chip_name`.

The new Registry.v defines lisp-esque association lists and a property
to ensure uniqueness of keys. Combined they create a `registry`.

Another potential application of a `registry` is ensuring pin names
aren't duplicated in a chip interface.